Secure automatic update
We now provide a secure and reliable way to retrieve updates from a server and
update the software without user intervention.

All releases are signed using a private certificate. The public key is embedded
in the binaries. The app downloads the release, check its checksum and its
signature. If the checksum matches and the signature is valid, the update is
installed and the app is restarted.

To assist developers and reducing the risk of error, go-task is used to
automate build. To do so, simply install `go-task` and run `task`. The default
command builds the app for all OS and arch. For each binary, it generates a
checksum file `*.sha256` and a signature file `*.sig`. Binaries are served from
a static file server.

The self update has only been tested on Linux.

The web app request the new server endpoint `/version` and shows the version
number to the user.